requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
deepl==1.18.0
distro==1.9.0
ecdsa==0.19.0
fastapi==0.112.2
geographiclib==2.0
geopy==2.4.1
googlemaps==4.10.0
groq==0.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jiter==0.5.0
openai==1.42.0
pyasn1==0.6.0
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.38.2
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
